+++ App/views/p_reply.py
@@ -4,6 +4,10 @@ from flask_jwt import jwt_required, current_identity
 
 from .index import index_views
 
+from App.controllers.p_comment import get_comment_by_id
+
+from App.controllers.user import is_admin
+
 from App.controllers.p_reply import (
     create_reply,
     get_all_replies_by_comment_id,
@@ -19,26 +23,35 @@ from App.controllers.p_reply import (
 reply_views = Blueprint("reply_views", __name__, template_folder="../templates")
 
 
-@reply_views.route("/product/comment/replies", methods=["GET"])
+@reply_views.route("/product/comment/<int:id>/reply", methods=["GET"])
 @jwt_required()
-def get_all_replies_by_comment_id_action():
-    data = request.json
-    replies = get_all_replies_by_comment_id_json(data["comment_id"])
+def get_all_replies_by_comment_id_action(id):
+    replies = get_all_replies_by_comment_id_json(id)
     if replies:
         return jsonify(replies), 200
     return jsonify([]), 200
 
 
-@reply_views.route("/product/comment/reply", methods=["POST"])
+@reply_views.route("/product/comment/reply/<int:id>", methods=["GET"])
 @jwt_required()
-def create_reply_action():
+def get_reply_by_id_action(id):
+    reply = get_reply_by_id_json(id)
+    if reply:
+        return jsonify(reply), 200
+    return jsonify({}), 404
+
+
+@reply_views.route("/product/comment/<int:id>/reply", methods=["POST"])
+@jwt_required()
+def create_reply_action(id):
+    comment = get_comment_by_id(id)
+    if not comment:
+        return jsonify({"message": "No comment found"}), 404
     data = request.json
-    create_reply(
-        comment_id=data["comment_id"],
-        user_id=data["user_id"],
-        body=data["body"]
+    reply = create_reply(
+        comment_id=id, user_id=current_identity.id, body=data["body"]
     )
-    return jsonify({"message": f"Reply {data['comment_id']} created"}), 201
+    return jsonify({"message": f"Reply {reply.id} created"}), 201
 
 
 @reply_views.route("/product/comment/reply/<int:id>", methods=["PUT"])
@@ -47,8 +60,10 @@ def update_reply_action(id):
     data = request.json
     reply = get_reply_by_id(id)
     if reply:
-        if 'body' in data:
-            update_reply(reply_id=id, body=data['body'])
+        if not reply.user_id == current_identity.id and not is_admin(current_identity):
+            return jsonify({"message": "Not authorized"}), 401
+        if "body" in data:
+            update_reply(reply_id=id, body=data["body"])
         return jsonify({"message": f"Reply {id} updated"}), 200
     return jsonify({"message": "No reply found"}), 404
 
@@ -58,6 +73,8 @@ def update_reply_action(id):
 def delete_reply_action(id):
     reply = get_reply_by_id(id)
     if reply:
+        if not reply.user_id == current_identity.id and not is_admin(current_identity):
+            return jsonify({"message": "Not authorized"}), 401
         delete_reply(id)
         return jsonify({"message": f"Reply {id} deleted"}), 200
     return jsonify({"message": "No reply found"}), 404
